+++ pkg/deployment/resources/inspector/secrets.go
@@ -29,6 +29,7 @@ import (
 
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/globals"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
 )
 
@@ -41,8 +42,8 @@ var secretsInspectorLoaderObj = secretsInspectorLoader{}
 type secretsInspectorLoader struct {
 }
 
-func (p secretsInspectorLoader) Component() throttle.Component {
-	return throttle.Secret
+func (p secretsInspectorLoader) Component() definitions.Component {
+	return definitions.Secret
 }
 
 func (p secretsInspectorLoader) Load(ctx context.Context, i *inspectorState) {