+++ pkg/deployment/resources/inspector/services.go
@@ -29,6 +29,7 @@ import (
 
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/globals"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
 )
 
@@ -41,8 +42,8 @@ var servicesInspectorLoaderObj = servicesInspectorLoader{}
 type servicesInspectorLoader struct {
 }
 
-func (p servicesInspectorLoader) Component() throttle.Component {
-	return throttle.Service
+func (p servicesInspectorLoader) Component() definitions.Component {
+	return definitions.Service
 }
 
 func (p servicesInspectorLoader) Load(ctx context.Context, i *inspectorState) {