+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Documentation) (1.3.0) EE & CE Definitions
 - (Improvement) Arango Kubernetes Client Mod Implementation
 - (Refactoring) Extract kerrors package
+- (Refactoring) Extract Inspector Definitions package
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress